[edarcarto] - add realm, retrofit, save user
