--- app/src/main/java/pe/edu/utp/yanapaapp/SingInActivity.java
@@ -1,32 +1,0 @@
-package pe.edu.utp.yanapaapp;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
-
-import androidx.annotation.Nullable;
-import androidx.appcompat.app.AppCompatActivity;
-
-public class SingInActivity extends AppCompatActivity {
-    Button btnLogin;
-    @Override
-    protected void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_sign_in);
-        btnLogin = (Button) findViewById(R.id.btnLogin);
-    }
-
-    @Override
-    protected void onResume() {
-        super.onResume();
-        btnLogin.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Intent i = new Intent(getApplicationContext(),MainActivity.class);
-                startActivity(i);
-                finish();
-            }
-        });
-    }
-}
