+++ app/src/main/java/pe/edu/utp/yanapaapp/activities/PrivacyActivity.java
